finish IMsAppData

+++ datum/flonum.hpp
@@ -37,9 +37,9 @@ namespace WarGrey::SCADA {
 	static const double pi        = 3.141592653589793;
 	static const long double pi_t = 3.1415926535897932385L;
 
-	static const double infinity_f  = std::numeric_limits<float>::infinity();
+	static const float infinity_f  = std::numeric_limits<float>::infinity();
 	static const double infinity    = std::numeric_limits<double>::infinity();
-	static const double infinity_t  = std::numeric_limits<long double>::infinity();
+	static const long double infinity_t  = std::numeric_limits<long double>::infinity();
 
 	bool inline flisnan(float f) { return std::isnan(f); }
 	double inline flisnan(double fl) { return std::isnan(fl); }
